Add Azurefile UT and separate volume tests
