+++ provider/aci_test.go
@@ -26,9 +26,6 @@ import (
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/intstr"
-
-	"github.com/golang/mock/gomock"
-	. "github.com/onsi/ginkgo"
 )
 
 const (
@@ -1065,406 +1062,5 @@ func TestCreatePodWithReadinessProbe(t *testing.T) {
 
 	if err := provider.CreatePod(context.Background(), pod); err != nil {
 		t.Fatal("Failed to create pod", err)
-	}
-}
-
-func TestCreatePodWithProjectedVolume(t *testing.T) {
-	podName := "pod-" + uuid.New().String()
-	podNamespace := "ns-name"
-	projectedVolumeName := "projectedvolume"
-	emptyVolumeName := "emptyVolumeName"
-	fakeSecretName := "fake-secret"
-	fakeShareName := "aksshare"
-	azureFileVolumeName := "azurefile"
-
-	aadServerMocker := NewAADMock()
-	aciServerMocker := NewACIMock()
-
-	aciServerMocker.OnGetRPManifest = func() (int, interface{}) {
-		manifest := &aci.ResourceProviderManifest{
-			Metadata: &aci.ResourceProviderMetadata{
-				GPURegionalSKUs: []*aci.GPURegionalSKU{
-					&aci.GPURegionalSKU{
-						Location: fakeRegion,
-						SKUs:     []aci.GPUSKU{aci.K80, aci.P100, aci.V100},
-					},
-				},
-			},
-		}
-
-		return http.StatusOK, manifest
-	}
-
-	mockCtrl := gomock.NewController(t)
-	podLister := NewMockPodLister(mockCtrl)
-	secretLister := NewMockSecretLister(mockCtrl)
-	configMapLister := NewMockConfigMapLister(mockCtrl)
-	serviceLister := NewMockServiceLister(mockCtrl)
-	pvcLister := NewMockPersistentVolumeClaimLister(mockCtrl)
-	pvLister := NewMockPersistentVolumeLister(mockCtrl)
-	mockSecretNamespaceLister := NewMockSecretNamespaceLister(mockCtrl)
-
-	encodedSecretVal := base64.StdEncoding.EncodeToString([]byte("fake-ca-data"))
-
-	aciServerMocker.OnCreate = func(subscription, resourceGroup, containerGroup string, cg *aci.ContainerGroup) (int, interface{}) {
-		assert.Check(t, is.Equal(fakeSubscription, subscription), "Subscription doesn't match")
-		assert.Check(t, is.Equal(fakeResourceGroup, resourceGroup), "Resource group doesn't match")
-		assert.Check(t, cg != nil, "Container group is nil")
-		assert.Check(t, is.Equal(podNamespace+"-"+podName, containerGroup), "Container group name is not expected")
-		assert.Check(t, cg.ContainerGroupProperties.Containers != nil, "Containers should not be nil")
-		assert.Check(t, is.Equal(1, len(cg.ContainerGroupProperties.Containers)), "1 Container is expected")
-		assert.Check(t, is.Equal("nginx", cg.ContainerGroupProperties.Containers[0].Name), "Container nginx is expected")
-		assert.Check(t, is.Equal(3, len(cg.Volumes)), "volume count not match")
-		assert.Check(t, is.Equal(emptyVolumeName, *cg.Volumes[0].Name), "volume name doesn't match")
-		assert.Check(t, is.Equal(azureFileVolumeName, *cg.Volumes[1].Name), "volume name doesn't match")
-		assert.Check(t, is.Equal(projectedVolumeName, *cg.Volumes[2].Name), "volume name doesn't match")
-		certVal := cg.Volumes[2].Secret["ca.crt"]
-		assert.Check(t, is.Equal(encodedSecretVal, *certVal), "configmap data doesn't match")
-
-		return http.StatusOK, cg
-	}
-
-	fakeSecret := &v1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      fakeSecretName,
-			Namespace: podNamespace,
-		},
-		Data: map[string][]byte{
-			azureFileStorageAccountName: []byte("azure storage account name"),
-			azureFileStorageAccountKey:  []byte("azure storage account key")},
-	}
-
-	fakeVolumeMount := v1.VolumeMount{
-		Name:      azureFileVolumeName,
-		MountPath: "/mnt/azure",
-	}
-
-	fakeVolumes := []v1.Volume{
-		{
-			Name: emptyVolumeName,
-			VolumeSource: v1.VolumeSource{
-				EmptyDir: &v1.EmptyDirVolumeSource{},
-			},
-		}, {
-			Name: azureFileVolumeName,
-			VolumeSource: v1.VolumeSource{
-				AzureFile: &v1.AzureFileVolumeSource{
-					ShareName:  fakeShareName,
-					SecretName: fakeSecretName,
-					ReadOnly:   true,
-				},
-			},
-		}, {
-			Name: projectedVolumeName,
-			VolumeSource: v1.VolumeSource{
-				Projected: &v1.ProjectedVolumeSource{
-					Sources: []v1.VolumeProjection{
-						{
-							ConfigMap: &v1.ConfigMapProjection{
-								LocalObjectReference: v1.LocalObjectReference{
-									Name: "kube-root-ca.crt",
-								},
-								Items: []v1.KeyToPath{
-									{
-										Key:  "ca.crt",
-										Path: "ca.crt",
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-
-	pod := &v1.Pod{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      podName,
-			Namespace: podNamespace,
-		},
-		Spec: v1.PodSpec{
-			Containers: []v1.Container{
-				v1.Container{
-					Name: "nginx",
-					ReadinessProbe: &v1.Probe{
-						Handler: v1.Handler{
-							HTTPGet: &v1.HTTPGetAction{
-								Port: intstr.FromInt(8080),
-								Path: "/",
-							},
-						},
-						InitialDelaySeconds: 10,
-						PeriodSeconds:       5,
-						TimeoutSeconds:      60,
-						SuccessThreshold:    3,
-						FailureThreshold:    5,
-					},
-				},
-			},
-		},
-	}
-	pod.Spec.Containers[0].VolumeMounts = append(pod.Spec.Containers[0].VolumeMounts, fakeVolumeMount)
-
-	for _, volume := range fakeVolumes {
-		if volume.AzureFile != nil {
-			secretLister.EXPECT().Secrets(podNamespace).Return(mockSecretNamespaceLister)
-			mockSecretNamespaceLister.EXPECT().Get(volume.AzureFile.SecretName).Return(fakeSecret, nil)
-		}
-	}
-
-	configMapNamespaceLister := NewMockConfigMapNamespaceLister(mockCtrl)
-	configMapLister.EXPECT().ConfigMaps(podNamespace).Return(configMapNamespaceLister)
-	configMapNamespaceLister.EXPECT().Get("kube-root-ca.crt").Return(&v1.ConfigMap{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: "kube-root-ca.crt",
-		},
-		Data: map[string]string{
-			"ca.crt": "fake-ca-data",
-			"foo":    "bar",
-		},
-	}, nil)
-
-	pod.Spec.Volumes = fakeVolumes
-
-	resourceManager, err := manager.NewResourceManager(podLister, secretLister, configMapLister, serviceLister, pvcLister, pvLister)
-	if err != nil {
-		t.Fatal("Unable to prepare the mocks for resourceManager", err)
-	}
-
-	provider, err := createTestProvider(aadServerMocker, aciServerMocker, resourceManager)
-	if err != nil {
-		t.Fatal("Unable to create test provider", err)
-	}
-
-	if err := provider.CreatePod(context.Background(), pod); err != nil {
-		t.Fatal("Failed to create pod", err)
-	}
-}
-
-func TestCreatePodWithCSIVolume(t *testing.T) {
-	podName := "pod-name"
-	podNamespace := "ns-name"
-	fakeVolumeSecret := "fake-volume-secret"
-	fakeShareName := "aksshare"
-	azureFileVolumeName := "azure"
-	emptyVolumeName := "emptyVolumeName"
-
-	aadServerMocker := NewAADMock()
-	aciServerMocker := NewACIMock()
-	mockCtrl := gomock.NewController(GinkgoT())
-	mockPodLister := NewMockPodLister(mockCtrl)
-	mockSecretLister := NewMockSecretLister(mockCtrl)
-	mockSecretNamespaceLister := NewMockSecretNamespaceLister(mockCtrl)
-	mockConfigMapLister := NewMockConfigMapLister(mockCtrl)
-	mockServiceLister := NewMockServiceLister(mockCtrl)
-	mockPvcLister := NewMockPersistentVolumeClaimLister(mockCtrl)
-	mockPvLister := NewMockPersistentVolumeLister(mockCtrl)
-
-	aciServerMocker.OnGetRPManifest = func() (int, interface{}) {
-		manifest := &aci.ResourceProviderManifest{
-			Metadata: &aci.ResourceProviderMetadata{
-				GPURegionalSKUs: []*aci.GPURegionalSKU{
-					{
-						Location: fakeRegion,
-						SKUs:     []aci.GPUSKU{aci.K80, aci.P100, aci.V100},
-					},
-				},
-			},
-		}
-		return http.StatusOK, manifest
-	}
-
-	aciServerMocker.OnCreate = func(subscription, resourceGroup, containerGroup string, cg *aci.ContainerGroup) (int, interface{}) {
-		assert.Check(t, is.Equal(fakeSubscription, subscription), "Subscription doesn't match")
-		assert.Check(t, is.Equal(fakeResourceGroup, resourceGroup), "Resource group doesn't match")
-		assert.Check(t, cg != nil, "Container group is nil")
-		assert.Check(t, is.Equal(podNamespace+"-"+podName, containerGroup), "Container group name is not expected")
-		assert.Check(t, cg.ContainerGroupProperties.Containers != nil, "Containers should not be nil")
-		assert.Check(t, is.Equal(1, len(cg.ContainerGroupProperties.Containers)), "1 Container is expected")
-		assert.Check(t, is.Equal("nginx", cg.ContainerGroupProperties.Containers[0].Name), "Container nginx is expected")
-		assert.Check(t, is.Equal(2, len(cg.Volumes)), "volume count not match")
-
-		return http.StatusOK, cg
-	}
-
-	configMapNamespaceLister := NewMockConfigMapNamespaceLister(mockCtrl)
-	mockConfigMapLister.EXPECT().ConfigMaps(podNamespace).Return(configMapNamespaceLister)
-	configMapNamespaceLister.EXPECT().Get("kube-root-ca.crt").Return(&v1.ConfigMap{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: "kube-root-ca.crt",
-		},
-		Data: map[string]string{
-			"ca.crt": "fake-ca-data",
-			"foo":    "bar",
-		},
-	}, nil)
-
-	fakeSecret := v1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      fakeVolumeSecret,
-			Namespace: podNamespace,
-		},
-		Data: map[string][]byte{
-			azureFileStorageAccountName: []byte("azure storage account name"),
-			azureFileStorageAccountKey:  []byte("azure storage account key")},
-	}
-
-	fakeVolumeMount := v1.VolumeMount{
-		Name:      azureFileVolumeName,
-		MountPath: "/mnt/azure",
-	}
-
-	fakePodVolumes := []v1.Volume{
-		{
-			Name: emptyVolumeName,
-			VolumeSource: v1.VolumeSource{
-				EmptyDir: &v1.EmptyDirVolumeSource{},
-			},
-		},
-		{
-			Name: azureFileVolumeName,
-			VolumeSource: v1.VolumeSource{
-				CSI: &v1.CSIVolumeSource{
-					Driver: "file.csi.azure.com",
-					VolumeAttributes: map[string]string{
-						azureFileSecretName: fakeVolumeSecret,
-						azureFileShareName:  fakeShareName,
-					},
-				},
-			},
-		},
-	}
-
-	cases := []struct {
-		description   string
-		secretVolume  *v1.Secret
-		volumes       []v1.Volume
-		expectedError error
-	}{
-		{
-			description:   "Secret is nil",
-			secretVolume:  nil,
-			volumes:       fakePodVolumes,
-			expectedError: fmt.Errorf("the secret %s for AzureFile CSI driver %s is not found", fakeSecret.Name, fakePodVolumes[1].Name),
-		},
-		{
-			description:   "Volume has a secret with a valid value",
-			secretVolume:  &fakeSecret,
-			volumes:       fakePodVolumes,
-			expectedError: nil,
-		},
-		{
-			description:  "Volume has no secret",
-			secretVolume: &fakeSecret,
-			volumes: []v1.Volume{{
-				Name: azureFileVolumeName,
-				VolumeSource: v1.VolumeSource{
-					CSI: &v1.CSIVolumeSource{
-						Driver:           "file.csi.azure.com",
-						VolumeAttributes: map[string]string{},
-					},
-				},
-			}},
-			expectedError: fmt.Errorf("secret volume attribute for AzureFile CSI driver %s cannot be empty or nil", azureFileVolumeName),
-		},
-		{
-			description:  "Volume has no share name",
-			secretVolume: &fakeSecret,
-			volumes: []v1.Volume{
-				{
-					Name: azureFileVolumeName,
-					VolumeSource: v1.VolumeSource{
-						CSI: &v1.CSIVolumeSource{
-							Driver: "file.csi.azure.com",
-							VolumeAttributes: map[string]string{
-								azureFileSecretName: fakeVolumeSecret,
-							},
-						},
-					},
-				}},
-			expectedError: fmt.Errorf("share name for AzureFile CSI driver %s cannot be empty or nil", fakePodVolumes[1].Name),
-		},
-		{
-			description:  "Volume is Disk Driver",
-			secretVolume: &fakeSecret,
-			volumes: []v1.Volume{
-				{
-					Name: azureFileVolumeName,
-					VolumeSource: v1.VolumeSource{
-						CSI: &v1.CSIVolumeSource{
-							Driver: "disk.csi.azure.com",
-							VolumeAttributes: map[string]string{
-								azureFileSecretName: fakeVolumeSecret,
-								azureFileShareName:  fakeShareName,
-							},
-						},
-					},
-				},
-			},
-			expectedError: fmt.Errorf("pod %s requires volume %s which is of an unsupported type %s", podName, azureFileVolumeName, "disk.csi.azure.com"),
-		},
-	}
-	for _, tc := range cases {
-		t.Run(tc.description, func(t *testing.T) {
-
-			resourceManager, err := manager.NewResourceManager(mockPodLister, mockSecretLister, mockConfigMapLister, mockServiceLister, mockPvcLister, mockPvLister)
-			if err != nil {
-				t.Fatal("Unable to prepare the mocks for resourceManager", err)
-			}
-
-			provider, err := createTestProvider(aadServerMocker, aciServerMocker, resourceManager)
-			if err != nil {
-				t.Fatal("Unable to create test provider", err)
-			}
-
-			pod := &v1.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      podName,
-					Namespace: podNamespace,
-				},
-				Spec: v1.PodSpec{
-					Containers: []v1.Container{
-						{
-							Name: "nginx",
-							ReadinessProbe: &v1.Probe{
-								Handler: v1.Handler{
-									HTTPGet: &v1.HTTPGetAction{
-										Port: intstr.FromInt(8080),
-										Path: "/",
-									},
-								},
-								InitialDelaySeconds: 10,
-								PeriodSeconds:       5,
-								TimeoutSeconds:      60,
-								SuccessThreshold:    3,
-								FailureThreshold:    5,
-							},
-						},
-					},
-				},
-			}
-
-			pod.Spec.Containers[0].VolumeMounts = append(pod.Spec.Containers[0].VolumeMounts, fakeVolumeMount)
-
-			for _, volume := range tc.volumes {
-				if volume.CSI != nil {
-					if volume.CSI.VolumeAttributes != nil && len(volume.CSI.VolumeAttributes) != 0 {
-						mockSecretLister.EXPECT().Secrets(podNamespace).Return(mockSecretNamespaceLister)
-						mockSecretNamespaceLister.EXPECT().Get(volume.CSI.VolumeAttributes[azureFileSecretName]).Return(tc.secretVolume, nil)
-					}
-				}
-			}
-
-			pod.Spec.Volumes = tc.volumes
-
-			err = provider.CreatePod(context.Background(), pod)
-
-			if tc.expectedError != nil {
-				assert.Equal(t, tc.expectedError.Error(), err.Error())
-			} else {
-				assert.NilError(t, tc.expectedError, err)
-			}
-		})
 	}
 }
